Consolidate upstream ports into phonenumberutil_test.go, matching upstream layout
The upstream test ports are the default now, so drop the redundant _upstream
suffix and mirror upstream's file organization (one file per test class, with
fixtures inline). Merges the four split port files plus the fixtures into a
single phonenumberutil_test.go, the analogue of upstream's PhoneNumberUtilTest.java.

Test layout now tracks upstream:
  PhoneNumberUtilTest.java         -> phonenumberutil_test.go
  ShortNumberInfoTest.java         -> shortnumberinfo_test.go (still ad-hoc, pending port)
  TestMetadataTestCase + RegionCode-> testmetadata_test.go (harness)
  (legacy real-metadata residue)   -> phonenumberutil_adhoc_test.go

No test changes; pure reorganization. Full suite passes (114 tests).
