--- phonenumberutil_upstream_test.go
@@ -1,100 +1,0 @@
-package phonenumbers
-
-// Faithful port of upstream libphonenumber's PhoneNumberUtilTest.java, run
-// against the synthetic test metadata (see testmetadata_test.go). Method names
-// and assertions mirror the Java tests so this file can be kept in sync with
-// upstream. Last reconciled against: v9.0.32
-//
-// This is an incremental migration: as a Java test is ported here, the
-// equivalent ad-hoc test in phonenumberutil_test.go (which runs against the
-// real embedded metadata and therefore breaks on metadata refreshes) should be
-// removed. Tests already present there are skipped here until that swap happens;
-// tests that need not-yet-implemented APIs (getSupportedTypesForRegion,
-// AsYouTypeFormatter, findNumbers, ...) are deferred to those features' PRs.
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestGetSupportedGlobalNetworkCallingCodes(t *testing.T) {
-	useTestMetadata(t)
-	globalNetworkCallingCodes := GetSupportedGlobalNetworkCallingCodes()
-	assert.NotEmpty(t, globalNetworkCallingCodes)
-	for callingCode := range globalNetworkCallingCodes {
-		assert.Greater(t, callingCode, 0)
-		assert.Equal(t, regionCode.UN001, GetRegionCodeForCountryCode(callingCode))
-	}
-}
-
-func TestGetSupportedCallingCodes(t *testing.T) {
-	useTestMetadata(t)
-	callingCodes := GetSupportedCallingCodes()
-	assert.NotEmpty(t, callingCodes)
-	for callingCode := range callingCodes {
-		assert.Greater(t, callingCode, 0)
-		assert.NotEqual(t, regionCode.ZZ, GetRegionCodeForCountryCode(callingCode))
-	}
-	// There should be more than just the global network calling codes in this set.
-	assert.Greater(t, len(callingCodes), len(GetSupportedGlobalNetworkCallingCodes()))
-	// But they should be included. Testing one of them.
-	assert.Contains(t, callingCodes, 979)
-}
-
-func TestGetInstanceLoadBadMetadata(t *testing.T) {
-	useTestMetadata(t)
-	assert.Nil(t, getMetadataForRegion("No Such Region"))
-	assert.Nil(t, getMetadataForNonGeographicalRegion(-1))
-}
-
-func TestGetInstanceLoadUSMetadata(t *testing.T) {
-	useTestMetadata(t)
-	metadata := getMetadataForRegion(regionCode.US)
-	assert.Equal(t, "US", metadata.GetId())
-	assert.Equal(t, int32(1), metadata.GetCountryCode())
-	assert.Equal(t, "011", metadata.GetInternationalPrefix())
-	assert.NotNil(t, metadata.NationalPrefix) // hasNationalPrefix
-	assert.Len(t, metadata.GetNumberFormat(), 2)
-	assert.Equal(t, `(\d{3})(\d{3})(\d{4})`, metadata.GetNumberFormat()[1].GetPattern())
-	assert.Equal(t, "$1 $2 $3", metadata.GetNumberFormat()[1].GetFormat())
-	assert.Equal(t, `[13-689]\d{9}|2[0-35-9]\d{8}`, metadata.GetGeneralDesc().GetNationalNumberPattern())
-	assert.Equal(t, `[13-689]\d{9}|2[0-35-9]\d{8}`, metadata.GetFixedLine().GetNationalNumberPattern())
-	assert.Len(t, metadata.GetGeneralDesc().GetPossibleLength(), 1)
-	assert.Equal(t, int32(10), metadata.GetGeneralDesc().GetPossibleLength()[0])
-	// Possible lengths are the same as the general description, so aren't stored
-	// separately in the toll free element as well.
-	assert.Len(t, metadata.GetTollFree().GetPossibleLength(), 0)
-	assert.Equal(t, `900\d{7}`, metadata.GetPremiumRate().GetNationalNumberPattern())
-	// No shared-cost data is available for US. Upstream leaves the descriptor
-	// unset (hasNationalNumberPattern() == false); our builder instead emits the
-	// legacy "NA" sentinel (matches nothing) for an absent descriptor — a known,
-	// functionally-equivalent divergence in metadata generation.
-	// TODO: align the builder's "NA" convention with upstream and restore
-	// assert.Empty here.
-	assert.Equal(t, "NA", metadata.GetSharedCost().GetNationalNumberPattern())
-}
-
-func TestGetInstanceLoadDEMetadata(t *testing.T) {
-	useTestMetadata(t)
-	metadata := getMetadataForRegion(regionCode.DE)
-	assert.Equal(t, "DE", metadata.GetId())
-	assert.Equal(t, int32(49), metadata.GetCountryCode())
-	assert.Equal(t, "00", metadata.GetInternationalPrefix())
-	assert.Equal(t, "0", metadata.GetNationalPrefix())
-	assert.Len(t, metadata.GetNumberFormat(), 6)
-	assert.Len(t, metadata.GetNumberFormat()[5].GetLeadingDigitsPattern(), 1)
-	assert.Equal(t, "900", metadata.GetNumberFormat()[5].GetLeadingDigitsPattern()[0])
-	assert.Equal(t, `(\d{3})(\d{3,4})(\d{4})`, metadata.GetNumberFormat()[5].GetPattern())
-	assert.Equal(t, "$1 $2 $3", metadata.GetNumberFormat()[5].GetFormat())
-	assert.Len(t, metadata.GetGeneralDesc().GetPossibleLengthLocalOnly(), 2)
-	assert.Len(t, metadata.GetGeneralDesc().GetPossibleLength(), 8)
-	// Nothing is present for fixed-line, since it is the same as the general
-	// desc, so for efficiency reasons we don't store an extra value.
-	assert.Len(t, metadata.GetFixedLine().GetPossibleLength(), 0)
-	assert.Len(t, metadata.GetMobile().GetPossibleLength(), 2)
-	assert.Equal(t, `(?:[24-6]\d{2}|3[03-9]\d|[789](?:0[2-9]|[1-9]\d))\d{1,8}`, metadata.GetFixedLine().GetNationalNumberPattern())
-	assert.Equal(t, "30123456", metadata.GetFixedLine().GetExampleNumber())
-	assert.Equal(t, int32(10), metadata.GetTollFree().GetPossibleLength()[0])
-	assert.Equal(t, `900([135]\d{6}|9\d{7})`, metadata.GetPremiumRate().GetNationalNumberPattern())
-}
